requirements.txt
streamlit
pandas
matplotlib
seaborn
mplsoccer
Pillow